+++ spec/javascripts/govuk_publishing_components/analytics-ga4/gtm-page-views.spec.js
@@ -9,29 +9,26 @@ describe('Google Tag Manager page view tracking', function () {
     saved.title = document.title
     document.title = 'This here page'
     expected = {
-      event: 'config_ready',
+      event: 'page_view',
       page: {
         location: document.location.href,
         referrer: document.referrer,
         title: 'This here page',
-        status_code: 200
-      },
-      publishing: {
+        status_code: '200',
+
         document_type: 'n/a',
         publishing_app: 'n/a',
         rendering_app: 'n/a',
         schema_name: 'n/a',
-        content_id: 'n/a'
-      },
-      taxonomy: {
+        content_id: 'n/a',
+
         section: 'n/a',
         taxon_slug: 'n/a',
         taxon_id: 'n/a',
         themes: 'n/a',
         taxon_slugs: 'n/a',
-        taxon_ids: 'n/a'
-      },
-      content: {
+        taxon_ids: 'n/a',
+
         language: 'n/a',
         history: 'false',
         withdrawn: 'false',
@@ -74,7 +71,7 @@ describe('Google Tag Manager page view tracking', function () {
 
   it('returns a page view with a specific status code', function () {
     window.httpStatusCode = 404
-    expected.page.status_code = 404
+    expected.page.status_code = '404'
     GOVUK.Gtm.sendPageView()
     expect(window.dataLayer[0]).toEqual(expected)
   })
@@ -111,7 +108,7 @@ describe('Google Tag Manager page view tracking', function () {
     for (var i = 0; i < tags.length; i++) {
       var tag = tags[i]
       createMetaTags(tag.tagName, tag.value)
-      expected.publishing[tag.gtmName] = tag.value
+      expected.page[tag.gtmName] = tag.value
     }
 
     GOVUK.Gtm.sendPageView()
@@ -155,7 +152,7 @@ describe('Google Tag Manager page view tracking', function () {
     for (var i = 0; i < tags.length; i++) {
       var tag = tags[i]
       createMetaTags(tag.tagName, tag.value)
-      expected.taxonomy[tag.gtmName] = tag.value
+      expected.page[tag.gtmName] = tag.value
     }
 
     GOVUK.Gtm.sendPageView()
@@ -165,7 +162,7 @@ describe('Google Tag Manager page view tracking', function () {
   it('returns a pageview with a language', function () {
     var html = document.querySelector('html')
     html.setAttribute('lang', 'wakandan')
-    expected.content.language = 'wakandan'
+    expected.page.language = 'wakandan'
 
     GOVUK.Gtm.sendPageView()
     expect(window.dataLayer[0]).toEqual(expected)
@@ -173,77 +170,77 @@ describe('Google Tag Manager page view tracking', function () {
 
   it('returns a pageview with history', function () {
     createMetaTags('content-has-history', 'true')
-    expected.content.history = 'true'
+    expected.page.history = 'true'
     GOVUK.Gtm.sendPageView()
     expect(window.dataLayer[0]).toEqual(expected)
   })
 
   it('returns a pageview without history', function () {
     createMetaTags('content-has-history', 'banana')
-    expected.content.history = 'false'
+    expected.page.history = 'false'
     GOVUK.Gtm.sendPageView()
     expect(window.dataLayer[0]).toEqual(expected)
   })
 
   it('returns a pageview on a withdrawn page', function () {
     createMetaTags('withdrawn', 'withdrawn')
-    expected.content.withdrawn = 'true'
+    expected.page.withdrawn = 'true'
     GOVUK.Gtm.sendPageView()
     expect(window.dataLayer[0]).toEqual(expected)
   })
 
   it('returns a pageview on a page with a first published date', function () {
     createMetaTags('first-published-at', '2022-03-28T19:11:00.000+00:00')
-    expected.content.first_published_at = '2022-03-28T19:11:00.000+00:00'
+    expected.page.first_published_at = '2022-03-28T19:11:00.000+00:00'
     GOVUK.Gtm.sendPageView()
     expect(window.dataLayer[0]).toEqual(expected)
   })
 
   it('returns a pageview on a page with a last updated date', function () {
     createMetaTags('updated-at', '2021-03-28T19:11:00.000+00:00')
-    expected.content.updated_at = '2021-03-28T19:11:00.000+00:00'
+    expected.page.updated_at = '2021-03-28T19:11:00.000+00:00'
     GOVUK.Gtm.sendPageView()
     expect(window.dataLayer[0]).toEqual(expected)
   })
 
   it('returns a pageview on a page with a last public updated date', function () {
     createMetaTags('public-updated-at', '2020-03-28T19:11:00.000+00:00')
-    expected.content.public_updated_at = '2020-03-28T19:11:00.000+00:00'
+    expected.page.public_updated_at = '2020-03-28T19:11:00.000+00:00'
     GOVUK.Gtm.sendPageView()
     expect(window.dataLayer[0]).toEqual(expected)
   })
 
   it('returns a pageview on a page marked with a publishing government', function () {
     createMetaTags('publishing-government', 'labour')
-    expected.content.publishing_government = 'labour'
+    expected.page.publishing_government = 'labour'
     GOVUK.Gtm.sendPageView()
     expect(window.dataLayer[0]).toEqual(expected)
   })
 
   it('returns a pageview on a page marked with a political status', function () {
     createMetaTags('political-status', 'ongoing')
-    expected.content.political_status = 'ongoing'
+    expected.page.political_status = 'ongoing'
     GOVUK.Gtm.sendPageView()
     expect(window.dataLayer[0]).toEqual(expected)
   })
 
   it('returns a pageview on a page marked with a primary publishing organisation', function () {
     createMetaTags('primary-publishing-organisation', 'Home Office')
-    expected.content.primary_publishing_organisation = 'Home Office'
+    expected.page.primary_publishing_organisation = 'Home Office'
     GOVUK.Gtm.sendPageView()
     expect(window.dataLayer[0]).toEqual(expected)
   })
 
   it('returns a pageview on a page marked with ids for contributing organisations', function () {
     createMetaTags('analytics:organisations', 'some organisations')
-    expected.content.organisations = 'some organisations'
+    expected.page.organisations = 'some organisations'
     GOVUK.Gtm.sendPageView()
     expect(window.dataLayer[0]).toEqual(expected)
   })
 
   it('returns a pageview on a page marked with world locations', function () {
     createMetaTags('analytics:world-locations', 'some world locations')
-    expected.content.world_locations = 'some world locations'
+    expected.page.world_locations = 'some world locations'
     GOVUK.Gtm.sendPageView()
     expect(window.dataLayer[0]).toEqual(expected)
   })